Initial Commit for YCSB-TS

+++ core/src/test/java/com/yahoo/ycsb/measurements/exporter/TestMeasurementsExporter.java
@@ -1,12 +1,12 @@
 /**
  * Copyright (c) 2015 Yahoo! Inc. All rights reserved.
- * <p>
+ * <p/>
  * Licensed under the Apache License, Version 2.0 (the "License"); you
  * may not use this file except in compliance with the License. You
  * may obtain a copy of the License at
- * <p>
+ * <p/>
  * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
+ * <p/>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
@@ -42,14 +42,14 @@ public class TestMeasurementsExporter {
         long max = 100000;
         ZipfianGenerator zipfian = new ZipfianGenerator(min, max);
         for (int i = 0; i < 1000; i++) {
-            int rnd = zipfian.nextInt();
+            long rnd = zipfian.nextLong();
             mm.measure("UPDATE", rnd);
         }
         mm.exportMeasurements(export);
         export.close();
 
         ObjectMapper mapper = new ObjectMapper();
-        JsonNode  json = mapper.readTree(out.toString("UTF-8"));
+        JsonNode json = mapper.readTree(out.toString("UTF-8"));
         assertTrue(json.isArray());
         assertEquals(json.get(0).get("measurement").asText(), "Operations");
         assertEquals(json.get(3).get("measurement").asText(), "MaxLatency(us)");
